Cleanup Manage Controller
