+++ BTCPayServer/Views/Manage/AddU2FDevice.cshtml
@@ -14,10 +14,11 @@
 
 
 <div class="card">
-    <h3 class="card-title">Performing U2F action</h3>
+    <h3 class="card-title px-1">Registering U2F Device</h3>
     <div class="card-body">
-        
         <p>You may be prompted to allow the site permission to access your security keys. After granting permission, the device will start to blink.</p>
+        
+        <a id="error-response" class="text-danger" href="javascript:window.location.reload()"> </a>
     </div>
     
     <div class="card-footer">
@@ -44,8 +45,7 @@
                 u2f.register(request.appId, registerRequests, [],
                     function(data) {
                         if (data.errorCode) {
-                            alert(errorMap[data.errorCode]);
-                            $("#btn-back").click();
+                            $("#error-response").text(errorMap[data.errorCode]);
                         }
                         console.log("Register callback", data);
                         $('#DeviceResponse').val(JSON.stringify(data));
